requirements.txt
Django>=3.1,<3.2
wagtail>=2.12,<2.13
django-widget-tweaks==1.4.8
psycopg2==2.8.6
gunicorn==20.0.4
django-heroku==0.3.1
whitenoise==5.2.0
wagtail-django-recaptcha==1.0
django-recaptcha==2.0.6
python-dotenv==0.15.0
django-storages==1.11.1
boto3==1.17.45 
botocore==1.20.45 
jmespath==0.10.0
python-dateutil==2.8.1 
s3transfer==0.3.6
django-sendgrid==v5-0.9.1 
future==0.18.2 
python-http-client==3.3.2 
sendgrid==6.6.0 
starkbank-ecdsa==1.1.0